requirements.txt
vobject
gunicorn==23.0.0
# juntagrico~=1.7.0
git+https://github.com/juntagrico/juntagrico@mag
juntagrico-pg~=1.7.0
juntagrico-godparent~=1.7.0
juntagrico-assignment-request~=1.7.0
# juntagrico-calendar~=1.6.0
git+https://github.com/mehalsgmues/juntagrico_calendar@mag
psycopg>=3.1.8
openpyxl==3.0.7
django-oauth-toolkit~=2.2.0
django-cors-headers==3.11.0
django-ckeditor==6.1.0
django-qr-code==2.2.0
django-multiselectfield==0.1.12
emoji==1.4.2
git+https://github.com/pascalfree/django-url-shortener
requests==2.32.5
django-admin-shell==2.0.1
IMAPClient==3.0.1